refactor: encapsulate scheduler timing methods with SchedulerTiming
